# api.yml
swagger: '2.0'
info:
    version: '0.1.0'
    title: 'auth'
host: 'localhost:8080'
#host: "auth--alberiw.repl.co"
schemes:
    - 'http'
    - 'https'
paths:
    /auth/login:
        post:
            tags:
                - 'auth'
            summary: ''
            description: ''
            operationId: 'login'
            consumes:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: ''
                  required: true
                  schema:
                      $ref: '#/definitions/Login'
            responses:
                200:
                    description: 'Ok'
                    schema:
                        $ref: '#/definitions/Token'
                500:
                    description: 'Internal Server Error'
    /auth/refresh:
        post:
            tags:
                - 'auth'
            summary: ''
            description: ''
            operationId: 'authRefresh'
            produces:
                - 'application/json'
            consumes:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: ''
                  required: true
                  schema:
                      type: 'string'
            responses:
                200:
                    description: 'Ok'
                    schema:
                        $ref: '#/definitions/Token'
                401:
                    description: 'Unathorized'
                500:
                    description: 'Internal Server Error'
    /auth/jwks:
        get:
            tags:
                - 'auth'
            summary: ''
            description: ''
            operationId: 'authJwks'
            produces:
                - 'application/json'
            responses:
                200:
                    description: 'Ok'
                    schema:
                        $ref: '#/definitions/Jwks'
                500:
                    description: 'Internal Server Error'
    /auth/logout:
        post:
            tags:
                - 'auth'
            summary: ''
            description: ''
            operationId: 'logout'
            consumes:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: ''
                  required: true
                  schema:
                      type: 'string'
            responses:
                200:
                    description: 'Ok'
                500:
                    description: 'Internal Server Error'
    /users:
        post:
            tags:
                - 'users'
            summary: ''
            description: ''
            operationId: 'createUser'
            consumes:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: ''
                  required: true
                  schema:
                      $ref: '#/definitions/CreateUserOption'
            responses:
                200:
                    description: 'Ok'
                500:
                    description: 'Internal Server Error'
    /users/{userId}:
        delete:
            tags:
                - 'users'
            summary: ''
            description: ''
            operationId: 'deleteUser'
            consumes:
                - 'application/json'
            parameters:
                - name: 'userId'
                  in: 'path'
                  description: ''
                  required: true
                  type: 'string'
                  format: 'uuid'
            responses:
                200:
                    description: 'Ok'
                401:
                    description: 'Unauthorized'
                500:
                    description: 'Internal Server Error'
            security:
                - jwt: []
    /users/{userId}/password:
        put:
            tags:
                - 'users'
            summary: ''
            description: ''
            operationId: 'changePassword'
            consumes:
                - 'application/json'
            parameters:
                - name: 'userId'
                  in: 'path'
                  description: ''
                  required: true
                  type: 'string'
                  format: 'uuid'
                - in: 'body'
                  name: 'body'
                  description: ''
                  required: true
                  schema:
                      $ref: '#/definitions/ChangePassword'
            responses:
                200:
                    description: 'Ok'
                401:
                    description: 'Unauthorized'
                500:
                    description: 'Internal Server Error'
            security:
                - jwt: []
securityDefinitions:
    jwt:
        type: 'apiKey'
        name: 'Authorization'
        in: 'header'
definitions:
    Login:
        type: 'object'
        properties:
            login:
                type: 'string'
            password:
                type: 'string'
        xml:
            name: 'Login'
    ChangePassword:
        type: 'object'
        properties:
            oldPassword:
                type: 'string'
            newPassword:
                type: 'string'
        xml:
            name: 'ChangePassword'
    CreateUserOption:
        type: 'object'
        properties:
            login:
                type: 'string'
            password:
                type: 'string'
        xml:
            name: 'User'
    Token:
        type: 'object'
        properties:
            access:
                type: 'string'
            refresh:
                type: 'string'
        xml:
            name: 'Token'
    Jwks:
        type: 'object'
        properties:
            keys:
                type: array
                items:
                    $ref: '#/definitions/Jwk'
        xml:
            name: 'Jwks'
    Jwk:
        type: 'object'
        properties:
            kid:
                type: 'string'
            use:
                type: 'string'
            alg:
                type: 'string'
            kty:
                type: 'string'
            n:
                type: 'string'
            e:
                type: 'string'
        xml:
            name: 'Jwk'